+++ contracts/testing/SmartContractLender.sol
@@ -16,7 +16,7 @@ contract SmartContractLender {
 
     function verifyLoanOffering(
         address[8],
-        uint[9],
+        uint256[9],
         uint32[2],
         bytes32
     )
@@ -30,7 +30,7 @@ contract SmartContractLender {
     function allow(
         address token,
         address spender,
-        uint amount
+        uint256 amount
     )
         external
         returns (bool _success)
